Color Vehicles and Warehouses by material
In Industry view, trucks and warehouses were colored by natural
resource. This made it hard to see detail of each supply chain.

This change colors them by the material they create. Eg, when Industry
Areas/Forestry view is activated, the colors are Logs, Lumber, Paper,
Planed Timber, Goods and Luxury Goods instead of merely Forestry. Only
assets involved in Forestry are highlighted in this view. This means that
if something is highlighted in any color, it is a participant in the
supply chain. Its color shows which stage of the supply chain it
participates in.